cleanup git service: it does not need authentication for core nodes any more
